kiwi: cpufreq: Average the load across affected cpus
We have two pairs of cores, each pair shares a clock and must run at the same
frequency.

Assuming that we have more tasks attempting to run on our cores than we have
cores, average the load over the cores instead of picking the max.

NOTE: "more tasks" is important.  If we have an equal number of runnable tasks
as we have cores then we still must pick the maximum load because one task
could be 100% cpu bound and the other only using 1% cpu and the only way to
accomodate that work-load is to go to 100% speed and have one wasted core.

This helps avoid unnecessary CPU spikes.  The effect is most noticeable when
you have a number of tasks that is not evenly divisible by the number of cores.
For example:

n_running 3 n_cpus 2 total_load 99 max_load 75 new_load 49

Here we have 3 tasks running on 2 cores and taking the maximum load is
actually over stating the performance needs that we have by 50% but when
averaged it is more sensible.

Change-Id: Iff1caef1bea9e1d6ff2e4102050062d6a3211ae1