cluster-plug: Treat offline little cores as "unloaded"
Prior to this change, if we bring the small cluster online and
externally it is brought offline on us, we will think it is online
but will not count the offline cores as idle.  This means we
will never think that we can take the cluster offline an therefore
will never return to a state where we can bring it online.

Treat offline little cores are being not busy.

Change-Id: I21fb506eb50c9297bbb7f3e40ac832bef017aeea